dashboard ui and ux
updated theme.dart with theme details,
moved bar and pie charts logic under new folder,
adapted home_screen per flutter analyze,
final ui and ux for dashboard screen metrics
